Fix the binding issue
